Interacción con datos

+++ app/Http/Controllers/EmpleadoController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Empleado;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 
 class EmpleadoController extends Controller
 {
@@ -79,9 +80,20 @@ class EmpleadoController extends Controller
      * @param  \App\Models\Empleado  $empleado
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Empleado $empleado)
+    public function update(Request $request, $id)
     {
-        //
+        $datosEmpleado = request()->except(['_token','_method']);
+
+        if($request->hasFile('Foto')){
+            $empleado=Empleado::findOrFail($id);
+            Storage::delete('public/'.$empleado->Foto);
+            $datosEmpleado['Foto']=$request->file('Foto')->store('uploads','public');
+        }
+
+        Empleado::where('id','=',$id)->update($datosEmpleado);
+        $empleado=Empleado::findOrFail($id);
+
+        return view('empleado.edit', compact('empleado'));
     }
 
     /**
@@ -92,7 +104,12 @@ class EmpleadoController extends Controller
      */
     public function destroy($id)
     {
-        Empleado::destroy($id);
+        $empleado=Empleado::findOrFail($id);
+        
+        if(Storage::delete('public/'.$empleado->Foto)){
+            Empleado::destroy($id);
+
+        }
 
         return redirect('empleado');
     }
